make a accordion and other fix
