+++ package.json
@@ -21,6 +21,7 @@
     "next": "14.0.0",
     "next-auth": "^4.24.5",
     "react": "^18",
+    "react-accessible-accordion": "^5.0.0",
     "react-detect-click-outside": "^1.1.7",
     "react-dom": "^18",
     "react-input-mask": "^2.0.4",
