+++ src/components/paramsFilter/index.module.scss
@@ -50,6 +50,7 @@
     background-color: white; /* or add it to the track */
   }
 }
+
 .disabled {
   transform: rotate(0deg);
   transition: all 0.5s ease-in-out;
